test(b): migrate commands/test_validate -> e2e/test_02_validate (#40)
Six CliRunner tests appended (now 8 total in test_02_validate.py).
TestValidate class dropped per Rule M2. cli_env fixture inlined via
project.write_config + project.copy_fixture + project.write_curation.
One test's monkeypatch of feather_etl.config.load_config preserved.
